+++ CHANGELOG.md
@@ -15,6 +15,7 @@ CentOS-6 6.8 x86_64, Apache 2.2, PHP 5.3, PHP memcached 1.0, PHP APC 3.1.
 - Updates `httpd` and `mod_ssl` packages.
 - Fixes noisy certificate generation output in logs during bootstrap when `APACHE_MOD_SSL_ENABLED` is `true`.
 - Changes `APACHE_SERVER_ALIAS` to a default empty value for `Makefile`, `scmi` and `systemd` templates which is the existing `Dockerfile` default.
+- Changes default `APACHE_SERVER_NAME` to unset and use the container's hostname for the Apache ServerName.
 
 ### 1.8.2 - 2017-01-24